+++ NEWS.md
@@ -1,5 +1,7 @@
 # vroom (development version)
 
+* `vroom()` now supports inputs with unnamed column types that are less than the number of columns (#296)
+
 * `vroom()` now outputs the correct column names even in the presence of skipped columns (#293, [tidyverse/readr#1215](https://github.com/tidyverse/readr/issues/1215))
 
 * `vroom_fwf(n_max=)` now works as intended when the input is a connection.